Fix win build for cred proxy
